Move KProveOptions to kprovex
