+++ kernel/src/main/java/org/kframework/kprove/KProveModule.java
@@ -3,10 +3,10 @@ package org.kframework.kprove;
 
 import com.google.inject.AbstractModule;
 import com.google.inject.Provides;
-import com.google.inject.TypeLiteral;
 import com.google.inject.multibindings.Multibinder;
 import com.google.inject.name.Named;
 import org.kframework.kompile.BackendModule;
+import org.kframework.kprovex.KProveOptions;
 import org.kframework.main.FrontEnd;
 import org.kframework.main.GlobalOptions;
 import org.kframework.main.Tool;
