+++ kernel/src/main/java/org/kframework/kprovex/KProveOptions.java
@@ -1,6 +1,6 @@
 // Copyright (c) K Team. All Rights Reserved.
 
-package org.kframework.kprove;
+package org.kframework.kprovex;
 
 import com.beust.jcommander.Parameter;
 import com.beust.jcommander.ParametersDelegate;
